+++ packages/engine/Source/Core/PolylineVolumeOutlineGeometry.js
@@ -83,7 +83,7 @@ function computeAttributes(positions, shape) {
  * @param {object} options Object with the following properties:
  * @param {Cartesian3[]} options.polylinePositions An array of positions that define the center of the polyline volume.
  * @param {Cartesian2[]} options.shapePositions An array of positions that define the shape to be extruded along the polyline
- * @param {Ellipsoid} [options.ellipsoid=Ellipsoid.WGS84] The ellipsoid to be used as a reference.
+ * @param {Ellipsoid} [options.ellipsoid=Ellipsoid.default] The ellipsoid to be used as a reference.
  * @param {number} [options.granularity=CesiumMath.RADIANS_PER_DEGREE] The distance, in radians, between each latitude and longitude. Determines the number of positions in the buffer.
  * @param {CornerType} [options.cornerType=CornerType.ROUNDED] Determines the style of the corners.
  *
@@ -124,7 +124,7 @@ function PolylineVolumeOutlineGeometry(options) {
   this._positions = positions;
   this._shape = shape;
   this._ellipsoid = Ellipsoid.clone(
-    defaultValue(options.ellipsoid, Ellipsoid.WGS84)
+    defaultValue(options.ellipsoid, Ellipsoid.default)
   );
   this._cornerType = defaultValue(options.cornerType, CornerType.ROUNDED);
   this._granularity = defaultValue(