+++ packages/engine/Source/Core/WebMercatorTilingScheme.js
@@ -13,8 +13,8 @@ import WebMercatorProjection from "./WebMercatorProjection.js";
  * @constructor
  *
  * @param {object} [options] Object with the following properties:
- * @param {Ellipsoid} [options.ellipsoid=Ellipsoid.WGS84] The ellipsoid whose surface is being tiled. Defaults to
- * the WGS84 ellipsoid.
+ * @param {Ellipsoid} [options.ellipsoid=Ellipsoid.default] The ellipsoid whose surface is being tiled. Defaults to
+ * the default ellipsoid.
  * @param {number} [options.numberOfLevelZeroTilesX=1] The number of tiles in the X direction at level zero of
  *        the tile tree.
  * @param {number} [options.numberOfLevelZeroTilesY=1] The number of tiles in the Y direction at level zero of
@@ -31,7 +31,7 @@ import WebMercatorProjection from "./WebMercatorProjection.js";
 function WebMercatorTilingScheme(options) {
   options = defaultValue(options, defaultValue.EMPTY_OBJECT);
 
-  this._ellipsoid = defaultValue(options.ellipsoid, Ellipsoid.WGS84);
+  this._ellipsoid = defaultValue(options.ellipsoid, Ellipsoid.default);
   this._numberOfLevelZeroTilesX = defaultValue(
     options.numberOfLevelZeroTilesX,
     1