+++ packages/engine/Source/Scene/SceneTransforms.js
@@ -4,6 +4,7 @@ import Cartesian3 from "../Core/Cartesian3.js";
 import Cartesian4 from "../Core/Cartesian4.js";
 import Cartographic from "../Core/Cartographic.js";
 import defined from "../Core/defined.js";
+import deprecationWarning from "../Core/deprecationWarning.js";
 import DeveloperError from "../Core/DeveloperError.js";
 import CesiumMath from "../Core/Math.js";
 import Matrix4 from "../Core/Matrix4.js";
@@ -27,6 +28,32 @@ const scratchWindowCoord0 = new Cartesian2();
 const scratchWindowCoord1 = new Cartesian2();
 
 /**
+ * Transforms a position in world (WGS84 or alternative ellipsoid) coordinates to window coordinates.  This is commonly used to place an
+ * HTML element at the same screen position as an object in the scene.
+ *
+ * @param {Scene} scene The scene.
+ * @param {Cartesian3} position The position in world (WGS84 or alternative ellipsoid) coordinates.
+ * @param {Cartesian2} [result] An optional object to return the input position transformed to window coordinates.
+ * @returns {Cartesian2} The modified result parameter or a new Cartesian2 instance if one was not provided.  This may be <code>undefined</code> if the input position is near the center of the ellipsoid.
+ *
+ * @example
+ * // Output the window position of longitude/latitude (0, 0) every time the mouse moves.
+ * const position = Cesium.Cartesian3.fromDegrees(0.0, 0.0);
+ * const handler = new Cesium.ScreenSpaceEventHandler(scene.canvas);
+ * handler.setInputAction(function(movement) {
+ *     console.log(Cesium.SceneTransforms.worldToWindowCoordinates(scene, position));
+ * }, Cesium.ScreenSpaceEventType.MOUSE_MOVE);
+ */
+SceneTransforms.worldToWindowCoordinates = function (scene, position, result) {
+  return SceneTransforms.worldWithEyeOffsetToWindowCoordinates(
+    scene,
+    position,
+    Cartesian3.ZERO,
+    result
+  );
+};
+
+/**
  * Transforms a position in WGS84 coordinates to window coordinates.  This is commonly used to place an
  * HTML element at the same screen position as an object in the scene.
  *
@@ -38,7 +65,7 @@ const scratchWindowCoord1 = new Cartesian2();
  * @example
  * // Output the window position of longitude/latitude (0, 0) every time the mouse moves.
  * const scene = widget.scene;
- * const ellipsoid = scene.globe.ellipsoid;
+ * const ellipsoid = scene.ellipsoid;
  * const position = Cesium.Cartesian3.fromDegrees(0.0, 0.0);
  * const handler = new Cesium.ScreenSpaceEventHandler(scene.canvas);
  * handler.setInputAction(function(movement) {
@@ -46,12 +73,11 @@ const scratchWindowCoord1 = new Cartesian2();
  * }, Cesium.ScreenSpaceEventType.MOUSE_MOVE);
  */
 SceneTransforms.wgs84ToWindowCoordinates = function (scene, position, result) {
-  return SceneTransforms.wgs84WithEyeOffsetToWindowCoordinates(
-    scene,
-    position,
-    Cartesian3.ZERO,
-    result
+  deprecationWarning(
+    "SceneTransforms.wgs84ToWindowCoordinates",
+    "SceneTransforms.wgs84ToWindowCoordinates has been deprecated. It will be removed in 1.21. Use SceneTransforms.worldToWindowCoordinates instead."
   );
+  return SceneTransforms.worldToWindowCoordinates(scene, position, result);
 };
 
 const scratchCartesian4 = new Cartesian4();
@@ -98,7 +124,7 @@ const scratchCameraPosition = new Cartesian3();
 /**
  * @private
  */
-SceneTransforms.wgs84WithEyeOffsetToWindowCoordinates = function (
+SceneTransforms.worldWithEyeOffsetToWindowCoordinates = function (
   scene,
   position,
   eyeOffset,
@@ -115,7 +141,7 @@ SceneTransforms.wgs84WithEyeOffsetToWindowCoordinates = function (
 
   // Transform for 3D, 2D, or Columbus view
   const frameState = scene.frameState;
-  const actualPosition = SceneTransforms.computeActualWgs84Position(
+  const actualPosition = SceneTransforms.computeActualEllipsoidPosition(
     frameState,
     position,
     actualPositionScratch
@@ -265,18 +291,46 @@ SceneTransforms.wgs84WithEyeOffsetToWindowCoordinates = function (
 };
 
 /**
- * Transforms a position in WGS84 coordinates to drawing buffer coordinates.  This may produce different
- * results from SceneTransforms.wgs84ToWindowCoordinates when the browser zoom is not 100%, or on high-DPI displays.
+ * Transforms a position in world coordinates to drawing buffer coordinates.  This may produce different
+ * results from SceneTransforms.worldToWindowCoordinates when the browser zoom is not 100%, or on high-DPI displays.
  *
  * @param {Scene} scene The scene.
- * @param {Cartesian3} position The position in WGS84 (world) coordinates.
+ * @param {Cartesian3} position The position in world (WGS84 or alternative ellipsoid) coordinates.
  * @param {Cartesian2} [result] An optional object to return the input position transformed to window coordinates.
  * @returns {Cartesian2} The modified result parameter or a new Cartesian2 instance if one was not provided.  This may be <code>undefined</code> if the input position is near the center of the ellipsoid.
  *
  * @example
  * // Output the window position of longitude/latitude (0, 0) every time the mouse moves.
- * const scene = widget.scene;
- * const ellipsoid = scene.globe.ellipsoid;
+ * const position = Cesium.Cartesian3.fromDegrees(0.0, 0.0);
+ * const handler = new Cesium.ScreenSpaceEventHandler(scene.canvas);
+ * handler.setInputAction(function(movement) {
+ *     console.log(Cesium.SceneTransforms.worldToDrawingBufferCoordinates(scene, position));
+ * }, Cesium.ScreenSpaceEventType.MOUSE_MOVE);
+ */
+SceneTransforms.worldToDrawingBufferCoordinates = function (
+  scene,
+  position,
+  result
+) {
+  result = SceneTransforms.worldToWindowCoordinates(scene, position, result);
+  if (!defined(result)) {
+    return undefined;
+  }
+
+  return SceneTransforms.transformWindowToDrawingBuffer(scene, result, result);
+};
+
+/**
+ * Transforms a position in world coordinates to drawing buffer coordinates.  This may produce different
+ * results from SceneTransforms.wgs84ToWindowCoordinates when the browser zoom is not 100%, or on high-DPI displays.
+ *
+ * @param {Scene} scene The scene.
+ * @param {Cartesian3} position The position in world (WGS84 or alternative ellipsoid) coordinates.
+ * @param {Cartesian2} [result] An optional object to return the input position transformed to window coordinates.
+ * @returns {Cartesian2} The modified result parameter or a new Cartesian2 instance if one was not provided.  This may be <code>undefined</code> if the input position is near the center of the ellipsoid.
+ *
+ * @example
+ * // Output the window position of longitude/latitude (0, 0) every time the mouse moves.
  * const position = Cesium.Cartesian3.fromDegrees(0.0, 0.0);
  * const handler = new Cesium.ScreenSpaceEventHandler(scene.canvas);
  * handler.setInputAction(function(movement) {
@@ -288,12 +342,16 @@ SceneTransforms.wgs84ToDrawingBufferCoordinates = function (
   position,
   result
 ) {
-  result = SceneTransforms.wgs84ToWindowCoordinates(scene, position, result);
-  if (!defined(result)) {
-    return undefined;
-  }
+  deprecationWarning(
+    "SceneTransforms.wgs84ToDrawingBufferCoordinates",
+    "SceneTransforms.wgs84ToDrawingBufferCoordinates has been deprecated. It will be removed in 1.21. Use SceneTransforms.worldToDrawingBufferCoordinates instead."
+  );
 
-  return SceneTransforms.transformWindowToDrawingBuffer(scene, result, result);
+  return SceneTransforms.worldToDrawingBufferCoordinates(
+    scene,
+    position,
+    result
+  );
 };
 
 const projectedPosition = new Cartesian3();
@@ -302,7 +360,7 @@ const positionInCartographic = new Cartographic();
 /**
  * @private
  */
-SceneTransforms.computeActualWgs84Position = function (
+SceneTransforms.computeActualEllipsoidPosition = function (
   frameState,
   position,
   result
@@ -398,7 +456,7 @@ const scratchWorldCoords = new Cartesian4();
 /**
  * @private
  */
-SceneTransforms.drawingBufferToWgs84Coordinates = function (
+SceneTransforms.drawingBufferToWorldCoordinates = function (
   scene,
   drawingBufferPosition,
   depth,